ENG-8641 Separate ExecutorVector from VoltDBEngine
